requirements.txt
requests
beautifulsoup4
hazm
python-dateutil
rapidfuzz
trafilatura